refactor: move Errors enum to util package
